fix compile warnings in VS

+++ vowpalwabbit/cost_sensitive.cc
@@ -143,7 +143,7 @@ void parse_label(parser* p, shared_data*sd, void* v, v_array<substring>& words)
       THROW(" invalid cost: specification -- no names on: " << words[i]);
 
     if (p->parse_name.size() == 1 || p->parse_name.size() == 2 || p->parse_name.size() == 3)
-    { f.class_index = sd->ldict ? sd->ldict->get(p->parse_name[0]) : (uint32_t)hashstring(p->parse_name[0], 0);
+    { f.class_index = sd->ldict ? (uint32_t)sd->ldict->get(p->parse_name[0]) : (uint32_t)hashstring(p->parse_name[0], 0);
       if (p->parse_name.size() == 1 && f.x >= 0)  // test examples are specified just by un-valued class #s
         f.x = FLT_MAX;
     }
