+++ vowpalwabbit/gd.h
@@ -28,7 +28,7 @@ struct multipredict_info { size_t count; size_t step; polyprediction* pred; regr
 inline void vec_add_multipredict(multipredict_info& mp, const float fx, uint64_t fi)
 { if ((-1e-10 < fx) && (fx < 1e-10)) return;
   weight*w    = mp.reg->weight_vector;
-  size_t mask = mp.reg->weight_mask;
+  uint64_t mask = mp.reg->weight_mask;
   polyprediction* p = mp.pred;
 
   fi &= mask;
@@ -48,7 +48,7 @@ inline void vec_add_multipredict(multipredict_info& mp, const float fx, uint64_t
 
 // iterate through one namespace (or its part), callback function T(some_data_R, feature_value_x, feature_weight)
 template <class R, void (*T)(R&, const float, float&)>
-inline void foreach_feature(weight* weight_vector, size_t weight_mask, features& fs, R& dat, uint64_t offset=0, float mult=1.)
+inline void foreach_feature(weight* weight_vector, uint64_t weight_mask, features& fs, R& dat, uint64_t offset=0, float mult=1.)
 {
   for (size_t i = 0; i < fs.size(); ++i)
     T(dat, mult*fs.values[i], weight_vector[(fs.indicies[i] + offset) & weight_mask]);
@@ -56,7 +56,7 @@ inline void foreach_feature(weight* weight_vector, size_t weight_mask, features&
 
 // iterate through one namespace (or its part), callback function T(some_data_R, feature_value_x, feature_index)
 template <class R, void (*T)(R&, float, uint64_t)>
-void foreach_feature(weight* /*weight_vector*/, size_t /*weight_mask*/, features& fs, R&dat, uint64_t offset=0, float mult=1.)
+void foreach_feature(weight* /*weight_vector*/, uint64_t /*weight_mask*/, features& fs, R&dat, uint64_t offset=0, float mult=1.)
 {
   for (size_t i = 0; i < fs.size(); ++i)
     T(dat, mult*fs.values[i], fs.indicies[i] + offset);
